# Improve Merkle Tree implementation