I first noticed that unbalanced implementation could panic if `MAX_N` was specified incorrectly.

I then looked closer into `active_levels` and it became clear that it is just a number of leaves, so decided to simplify the implementation. Especially balanced implementation is very elegant now.

There are no dedicated Merkle Tree benchmarks yet, but archiving that uses it shows clear and non-negligible improvement.

```
Before:
segment-archiving-whole-segment
                        time:   [1.9043 s 1.9133 s 1.9230 s]
After:
segment-archiving-whole-segment
                        time:   [1.8101 s 1.8152 s 1.8202 s]
```